correção de arquitetura
