+++ jesus-king-tech/src/main/java/com/api/jesus_king_tech/domain/produto/repository/ProdutoRepository.java
@@ -1,5 +1,6 @@
-package com.api.jesus_king_tech.produto;
+package com.api.jesus_king_tech.domain.produto.repository;
 
+import com.api.jesus_king_tech.domain.produto.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
