+++ jesus-king-tech/src/main/java/com/api/jesus_king_tech/domain/produto/tipo/repository/TipoRepository.java
@@ -1,8 +1,7 @@
-package com.api.jesus_king_tech.produto;
+package com.api.jesus_king_tech.domain.produto.tipo.repository;
 
+import com.api.jesus_king_tech.domain.produto.tipo.Tipo;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface TipoRepository extends JpaRepository<Tipo, Integer> {
