--- jesus-king-tech/src/main/java/com/api/jesus_king_tech/produto/CategoriaDTO.java
@@ -1,29 +1,0 @@
-package com.api.jesus_king_tech.produto;
-
-import java.util.List;
-
-public class CategoriaDTO {
-
-    private Integer id;
-    private String nome;
-
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-
-}
-
