+++ jesus-king-tech/src/main/java/com/api/jesus_king_tech/service/ProdutoService.java
@@ -1,0 +1,5 @@
+package com.api.jesus_king_tech.service;
+
+public class ProdutoService {
+
+}
